refactor(qc): rename qc_booleans package -> qc_flags
Rename kpfpipe/quality_control/qc_booleans/ -> qc_flags/ (and
tests/regression/test_qc_booleans.py -> test_qc_flags.py), updating all imports
(__init__, level0/1/2, recipes/kpf_drp_science.py, scripts/qc.py) and the
profiling attribution string in tests/profiling/_profiling.py.

Also correct four "qc_booleans validator" prose mentions (data_models/base.py,
keyword_registry.py, CLAUDE.md, a test docstring) that were doubly stale -- the
header validator moved to the checkpoints layer in the prior refactor -- to
"checkpoints validator" / quality_control/checkpoints/base.py.

No behavior change.
