--- kpfpipe/quality_control/qc_booleans/__init__.py
@@ -1,6 +1,0 @@
-from kpfpipe.quality_control.qc_booleans.base import QC
-from kpfpipe.quality_control.qc_booleans.level0 import QCL0
-from kpfpipe.quality_control.qc_booleans.level1 import QCL1
-from kpfpipe.quality_control.qc_booleans.level2 import QCL2
-
-__all__ = ["QC", "QCL0", "QCL1", "QCL2"]
