+++ kpfpipe/quality_control/qc_flags/__init__.py
@@ -1,0 +1,6 @@
+from kpfpipe.quality_control.qc_flags.base import QC
+from kpfpipe.quality_control.qc_flags.level0 import QCL0
+from kpfpipe.quality_control.qc_flags.level1 import QCL1
+from kpfpipe.quality_control.qc_flags.level2 import QCL2
+
+__all__ = ["QC", "QCL0", "QCL1", "QCL2"]
